--- app/examples/circleci/configure-docker-credentials-store.yml
@@ -1,40 +1,0 @@
-version: 2.1
-
-orbs:
-  docker: circleci/docker@2.4.0
-
-commands:
-  install_deps:
-    description: "Install required dependencies"
-    steps:
-      - run:
-          name: Install Dependencies
-          command: |
-            apt-get update
-            apt-get install -y curl jq
-
-jobs:
-  configure-docker-credentials-store:
-    docker:
-      - image: cimg/base:stable
-    steps:
-      - checkout
-      - install_deps
-      - docker/configure-docker-credentials-store
-
-  configure-docker-credentials-store-custom:
-    docker:
-      - image: cimg/base:stable
-    steps:
-      - checkout
-      - install_deps
-      - docker/configure-docker-credentials-store:
-          docker-config-path: /custom/path/.docker/config.json
-          helper-name: pass
-
-workflows:
-  version: 2
-  configure-docker-credentials:
-    jobs:
-      - configure-docker-credentials-store
-      - configure-docker-credentials-store-custom
