+++ app/examples/circleci/credentials.yml
@@ -14,6 +14,24 @@ commands:
             apt-get install -y curl jq git gnupg2 pass
 
 jobs:
+  configure-docker-credentials-store:
+    docker:
+      - image: cimg/base:stable
+    steps:
+      - checkout
+      - install_deps
+      - docker/configure-docker-credentials-store
+
+  configure-docker-credentials-store-custom:
+    docker:
+      - image: cimg/base:stable
+    steps:
+      - checkout
+      - install_deps
+      - docker/configure-docker-credentials-store:
+          docker-config-path: /custom/path/.docker/config.json
+          helper-name: pass
+
   install:
     docker:
       - image: cimg/base:stable
@@ -35,6 +53,11 @@ jobs:
 
 workflows:
   version: 2
+  configure-docker-credentials:
+    jobs:
+      - configure-docker-credentials-store
+      - configure-docker-credentials-store-custom
+  
   install-and-build:
     jobs:
       - install
